main/requirements.txt
fastapi==0.115.5
fastapi-cache2==0.2.2
python-dotenv==1.0.1
uvicorn==0.32.1

